feat: add scripts from testnet upgrade v3
 8 -> register operators to bootstrap contract
 9 -> edit bootstrap time (DEPENDS on script content!)
10 -> deploy only exocore gateway on our chain
11 -> set peers on both sides (I couldn't get markBootstrap to work here
so I had to `cast send` the transaction manually
12 -> redeploys the (modified) client chain gateway. since the previous
upgrade transaction failed due to an out of gas error, the client chain
gateway logic was modified to remove the loops during initialization,
and redeployed. this script also calls `setClientChainGatewayLogic`
appropriately.

Also included are the contract addresses as JSON for each deployment
script above.