build(ci): merge workflows and use cache
We now have a multi-job workflow, which means it will run the jobs in
parallel. Between jobs, we cache the artifacts generated by Foundry so
they can be reused. The artifacts cached are the result of compilation
by Foundry, and hence, they can be reused across jobs. The `build` job
is responsible for generating them, while the `test` and `fmt` jobs can
reuse them, which is why they `need` the `build` job.