rename operator to validator in bootstrap
We are using, in documentation as well in code, a distinction between
the role of an operator and that of a validator. The set of validators
is a subset of the set of operators; in other words, every validator is
considered as an operator. However, not every operator is a validator.

A validator is an entity registered to validate the Exocore chain. An
operator is an entity registered to validate any AVS, including the
dogfood AVS which represents the Exocore chain.

Before the network is considered bootstrapped, the only way to register
as an operator is to register as a validator as well; hence, we rename
within the (exclusively) Bootstrap context `operator` as `validator`.
Any entity registering itself as a validator in the Bootstrap contract
will automatically get registered as an operator when the chain is
started, as a consequence.