fix(ci): downgrade node due to nodejs/node#53902 (#56)
A bug in node version 22.5.0 has caused the CI to fail. While a patch
has been merged, a release is still awaited. This PR works around that
by pinning the node version to one that does not contain the bug.

Once there is a new release, the pinned part (`.4`) can be removed