refactor: add natspec documentation (#58)
* rename operator to validator in bootstrap

We are using, in documentation as well in code, a distinction between
the role of an operator and that of a validator. The set of validators
is a subset of the set of operators; in other words, every validator is
considered as an operator. However, not every operator is a validator.

A validator is an entity registered to validate the Exocore chain. An
operator is an entity registered to validate any AVS, including the
dogfood AVS which represents the Exocore chain.

Before the network is considered bootstrapped, the only way to register
as an operator is to register as a validator as well; hence, we rename
within the (exclusively) Bootstrap context `operator` as `validator`.
Any entity registering itself as a validator in the Bootstrap contract
will automatically get registered as an operator when the chain is
started, as a consequence.

* refactor(all): add natspec docs to src

* chore(ci): remove extra space

* fix(ci): move or add slither warning

* fix(header): add MIT license

* fix: rename file from operator to validator

* fix: merkleize typo

* doc: remove unused function doc

* doc: update doc for new design

* Apply suggestions from code review

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* fix: remove references to validator set

* Update docs/client-chain-contracts-design.md

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* fix: rename param for clarity

---------

Co-authored-by: call-by <smart.solidity@gmail.com>
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>