refactor(precompiles): chain as source of truth
The `ExocoreGateway` allows the contract owner to call precompiles which
allow the registration of a client chain or a token. It also enables
updated of these items. To check whether a request is an update or an
addition, it relies on the ExocoreGateway keeping a copy of registered
items. However, this does not work for items which were registered at
genesis.

This PR works around that problem by requiring that the single source of
truth be the precompiles, whose responsibility is to report if a request
was an update or an addition.