refactor(precompiles): chain as source of truth (#66)
* refactor(precompiles): chain as source of truth

The `ExocoreGateway` allows the contract owner to call precompiles which
allow the registration of a client chain or a token. It also enables
updated of these items. To check whether a request is an update or an
addition, it relies on the ExocoreGateway keeping a copy of registered
items. However, this does not work for items which were registered at
genesis.

This PR works around that problem by requiring that the single source of
truth be the precompiles, whose responsibility is to report if a request
was an update or an addition.

* fix(precompile): pluralize token to tokens

* fix: check client chain before setPeer

* fix(doc): clarify comment

* refactor(gateway): use IAssets directly not `call`

After the precompiles for assets and client chains were merged, the
convoluted mechanism of calling the `getClientChains` function via the
selector is not necessary.

* fix(test): expect events