fix: use `block.timestamp >= lockTime`
The `isLocked` function returns true if `block.timestamp >= lockTime`,
which is exactly what the `_validateSpawnTimeAndOffsetDuration` function
should do.

The spawn time must not be in the past, or exactly at the current block.

The offset duration must be greater than equal to the spawn time,
however, if they are equal, the lock time ends up in the past, and it is
rejected.