feat(tvl-limit): impose tvl limit on client chain
We can consider the TVL limits to be of two types: one is the logical
TVL limit which disallows deposits from exceeding total supply and the
other is the operational TVL limit which is used for risk management
during the launch phase of the network.

The former is imposed by Exocore's precompiles while the latter is
imposed by the client chain's gateway contracts. Both fields are
editable by the contract owners: the total supply in response to minting
and burning and the deposit limits with time to larger values.

As a consequence of the introduction of a limit on the client chain, it
must now be provided at the time of token registration. It can be
updated or reduced unilaterally on the client chain.

While editing the `totalSupply` on Exocore, it is best to be fully
accurate. Reducing the `totalSupply` to lower than the TVL limit
configured on the client chain will result in failed deposits which may
produce a system halt. Even if the TVL limit is reduced on the client
chain after this, it can impact the messages that are already in-flight.