fix: simplify TVL limit and total supply limit
Based on the discussion, it was decided to remove the total supply limit
enforced by Exocore. Instead, the TVL limit will be imposed on the
client chain, which will also enforce that any transfers more than the
total supply will fail.

It is possible to set a value of infinite TVL limit by using
`type(uin256).max`