add testnet v5 deployment addresses (#95)
* add testnet v5 deployment addresses

* fix(script): easier SetPeer call

It is a known fact that Foundry does not play well with custom
precompiles. In some cases, even passing `--skip-simulation` does not
satisfy its tastes.

This change attempts to fix one of those cases by having the `IAssets`
mock return a value of `true` for the `clientChainId`. This is done by
deploying the `AssetsMock` contract with an appropriate constructor call
to the precompile address.

Note that this fix applies only if USE_EXOCORE_PRECOMPILE_MOCK=false and
not otherwise.