refactor: use correct constants
A lot of the parameters such as the beacon genesis time, the deneb time
and the deposit address are dependent on the network chosen for
deployment. This change allows a library to provide these constants,
based on the `block.chainid`.

Additionally, it creates the possibility of using a separately deployed
contract for integration networks (refer to the next commit), for which,
the parameters can be dynamically provided via `INetworkConfig`.

As a side effect, the default chain_id used in tests has been changed to
let `NetworkConstants` library provide these constants during tests.