fix: use the validatorID instead of pubkey
On an Ethereum-based client chain, a validator container (generated as
part of the proof) does not have the raw public key stored within it.
Rather, it is the hash of the public key. This hash, therefore, cannot
be used to query the beacon chain API.

This PR proposes an alternative solution of using the `validatorIndex`,
that is, the position of the validator amongst all of the validators, be
used instead. Note that, according to the beacon spec, this number does
not change even as more validators enter or leave the network.

The `validatorIndex` is passed to the ExocoreGateway as a unique
identifier for it to send to the precompile, which refers to it as the
`validatorID`. For other chains, where a validator public key may be
directly available, the `validatorID` will be the public key instead of
the index. On Ethereum, the public key is 48 bytes (though we pass a
uint256 index which is 32 bytes). On Solana, the public key is 32 bytes
and on Sui it is 96 bytes. To handle the varying lengths, the public key
has been moved to the end of the payload, such that, it is the remainder
of the payload after parsing the staker and the amount.

BREAKING CHANGE
Since the PR modifies the format of the data sent over the wire, it is
not compatible with prior deployments of ClientChainGateway and
ExocoreGateway. If care is taken to ensure that there are no affected
messages currently in-flight and the implementations are updated in
tandem, it will be sufficient.