fix(generate): support whitelist `nstETH` + !`ETH`
In the event that the Bootstrap contract has whitelisted `nstETH` but
not `ETH`, the `nstETH` entry will only fetch the effective balance and
the token price from the `ETH` entry will be missing. This commit fixes
that by manually adding an `ETH` entry to the oracle state in such an
event.