feat(ci): carve out storage layout CI (#127)
* feat(ci): carve out storage layout CI

Since the storage layout CI needs the API keys to run, it does not work
in PRs originating from forked repos. To avoid those unnecessary errors,
this PR splits the workflow into two parts: the first that generates the
compiled layouts and the second that fetches the deployed layouts and
compares them.

Such a split is possible because the second type of workflow runs in a
trusted context, and hence, the API key is accessible to it.

* fix(ci): update status fast

* feat(ci): carve out storage layout CI

Since the storage layout CI needs the API keys to run, it does not work
in PRs originating from forked repos. To avoid those unnecessary errors,
this PR splits the workflow into two parts: the first that generates the
compiled layouts and the second that fetches the deployed layouts and
compares them.

Such a split is possible because the second type of workflow runs in a
trusted context, and hence, the API key is accessible to it.

* respond to AI comments

* respond to more AI comments

* error out if expected files do not exist

* fix(ci): escape the params

* fix(ci): fail faster if the parent job failed

* rename input for clarity

* do not set the key unless required

* use not condition instead of == false

* remove superfluous comment

* fix(ci): add timeout

* feat(forge-ci): parallelize layout extraction

* fix(ci): use artifacts with different names

* fix(ci): flatten combination correctly

* split deployed layout job to parallel

* fix(ci): make matrix single line output

* fix(ci): download current run artifact

* fix(ci): remove debug

* fix(ci): always create matrix

* fix(ci): update job timeouts

* fix(ci): respond to AI comments