feat(generate): native restaking genesis (#123)
* refactor: use correct constants

A lot of the parameters such as the beacon genesis time, the deneb time
and the deposit address are dependent on the network chosen for
deployment. This change allows a library to provide these constants,
based on the `block.chainid`.

Additionally, it creates the possibility of using a separately deployed
contract for integration networks (refer to the next commit), for which,
the parameters can be dynamically provided via `INetworkConfig`.

As a side effect, the default chain_id used in tests has been changed to
let `NetworkConstants` library provide these constants during tests.

* feat: introduce dynamic `NetworkConfig`

* attempt to add deposit and prove scripts

* fix: update proof generation script + generate.js

* change workflow cause

* Revert "change workflow cause"

This reverts commit 183dfabbc42addad7285f903680a7dc073294e8a.

* fix: respond to AI comments

* fix: respond to more AI comments

* fix(integration): remove container.json check

* fix(env): clarify port usage

* fix(generate): respond to review comments

* fix(generate): update x/oracle data

- The `nstETH` token needs to be added both as a token and as a feeder
- The `nstETH` token's `asset_id` needs to map back to the native
  (non-staked) token's `asset_id`

* fix(generate): support whitelist `nstETH` + !`ETH`

In the event that the Bootstrap contract has whitelisted `nstETH` but
not `ETH`, the `nstETH` entry will only fetch the effective balance and
the token price from the `ETH` entry will be missing. This commit fixes
that by manually adding an `ETH` entry to the oracle state in such an
event.

* update for effective balance

* respond to AI comments

* update for 31 ETH test

* fix storage layout

* fix(integration): add client chain gateway

* fix(integration): make Bootstrap address constant

* fix(generate): truncate the values

* fix: make bootstrapper

* fix: add validation for deposit address (ai)